Added support for proxy engines
